+++ app.go
@@ -449,7 +449,7 @@ func (a *App) GetAppInfo() map[string]string {
 	version, _ := a.encoder.GetFFmpegVersion()
 	return map[string]string{
 		"appName":        "SyncLauper VideoConverter",
-		"appVersion":     "1.2.0",
+		"appVersion":     "1.2.1",
 		"ffmpegVersion":  version,
 	}
 }